python: Move Duration and Timestamp out of codegen
