+++ python/foxglove-sdk/python/foxglove/_foxglove_py/schemas_wkt.pyi
@@ -1,0 +1,23 @@
+from typing import Optional
+
+class Duration:
+    """
+    A duration in seconds and nanoseconds
+    """
+
+    def __new__(
+        cls,
+        sec: int,
+        nsec: Optional[int] = None,
+    ) -> "Duration": ...
+
+class Timestamp:
+    """
+    A timestamp in seconds and nanoseconds
+    """
+
+    def __new__(
+        cls,
+        sec: int,
+        nsec: Optional[int] = None,
+    ) -> "Timestamp": ...
