+++ python/foxglove-sdk/src/lib.rs
@@ -20,6 +20,7 @@ use websocket_server::{
 
 mod errors;
 mod generated;
+mod schemas_wkt;
 mod websocket_server;
 
 #[pyclass(module = "foxglove")]
