+++ typescript/schemas/src/internal/generatePyclass.ts
@@ -21,7 +21,11 @@ export function generateSchemaPrelude(): string {
     `#![allow(non_snake_case)]`,
   ];
 
-  const imports = ["use pyo3::{prelude::*, types::PyBytes};"];
+  const imports = [
+    "use pyo3::prelude::*;",
+    "use pyo3::types::PyBytes;",
+    "use crate::schemas_wkt::{Duration, Timestamp};",
+  ];
 
   const outputSections = [docs.join("\n"), imports.join("\n")];
 
@@ -44,8 +48,6 @@ export function generatePySchemaStub(schemas: FoxgloveSchema[]): string {
     "from enum import Enum",
     "from typing import List, Optional",
   ].join("\n") + "\n";
-
-  const timeTypes = generateTimeTypeStubs();
 
   const enums = schemas
     .filter((schema) => schema.type === "enum")
@@ -86,7 +88,7 @@ export function generatePySchemaStub(schemas: FoxgloveSchema[]): string {
 
   // Enums come first to provide default values for constructor parameters. Otherwise, sort by name.
   const enumSources = enums.sort((a, b) => a.name.localeCompare(b.name)).map(({ source }) => source);
-  const classSources = [...classes, ...timeTypes]
+  const classSources = classes
     .sort((a, b) => a.name.localeCompare(b.name))
     .map(({ source }) => source);
 
@@ -450,102 +452,6 @@ function structName(name: string): string {
     return "GeoJson";
   }
   return name;
-}
-
-/**
- * .pyi stubs for Timestamp and Duration.
- */
-function generateTimeTypeStubs(): { name: string, source: string }[] {
-  const timestamp = `
-class Timestamp:
-    """
-    A timestamp in seconds and nanoseconds
-    """
-    def __new__(
-        cls,
-        sec: int,
-        nsec: Optional[int] = None,
-    ) -> "Timestamp": ...
-`;
-
-  const duration = `
-class Duration:
-    """
-    A duration in seconds and nanoseconds
-    """
-    def __new__(
-        cls,
-        sec: int,
-        nsec: Optional[int] = None,
-    ) -> "Duration": ...
-`;
-
-  return [{ name: "Timestamp", source: timestamp }, { name: "Duration", source: duration }];
-}
-
-/**
- * Defines a struct for representing Timestamp and Duration.
- *
- * This also provides a `From` implementation into prost types for proto serialization.
- */
-export function generateTimeTypes(): string {
-  return `
-/// A timestamp in seconds and nanoseconds
-///
-/// :param sec: The number of seconds since a user-defined epoch.
-/// :param nsec: The number of nanoseconds since the :py:attr:\`sec\` value.
-#[pyclass(module = "foxglove.schemas")]
-#[derive(Clone)]
-pub struct Timestamp(pub(crate) foxglove::schemas::Timestamp);
-
-#[pymethods]
-impl Timestamp {
-    #[new]
-    #[pyo3(signature = (sec, nsec=None))]
-    fn new(sec: u32, nsec: Option<u32>) -> Self {
-        let nsec = nsec.unwrap_or(0);
-        Self(foxglove::schemas::Timestamp{ sec, nsec })
-    }
-
-    fn __repr__(&self) -> String {
-        format!("Timestamp(sec={}, nsec={})", self.0.sec, self.0.nsec).to_string()
-    }
-}
-
-impl From<Timestamp> for foxglove::schemas::Timestamp {
-    fn from(value: Timestamp) -> Self {
-        value.0
-    }
-}
-
-/// A duration, composed of seconds and nanoseconds
-///
-/// :param sec: The number of seconds in the duration.
-/// :param nsec: The number of nanoseconds in the positive direction.
-#[pyclass(module = "foxglove.schemas")]
-#[derive(Clone)]
-pub struct Duration(pub(crate) foxglove::schemas::Duration);
-
-#[pymethods]
-impl Duration {
-    #[new]
-    #[pyo3(signature = (sec, nsec=None))]
-    fn new(sec: i32, nsec: Option<u32>) -> Self {
-        let nsec = nsec.unwrap_or(0);
-        Self(foxglove::schemas::Duration{ sec, nsec })
-    }
-
-    fn __repr__(&self) -> String {
-      format!("Duration(sec={}, nsec={})", self.0.sec, self.0.nsec).to_string()
-    }
-}
-
-impl From<Duration> for foxglove::schemas::Duration {
-    fn from(value: Duration) -> Self {
-        value.0
-    }
-}
-`;
 }
 
 /**
